* reference and tunable class are separated in extra files * non-atomic reduction is not working after merge, some problems with parameters synchronization
